refactor: added MidiViewModel and MidiView
