+++ src/Spice86/Views/MidiView.axaml.cs
@@ -1,0 +1,8 @@
+namespace Spice86.Views;
+using Avalonia.Controls;
+
+public partial class MidiView : UserControl {
+    public MidiView() {
+        InitializeComponent();
+    }
+}
